requirements.txt
importlib_resources; python_version<'3.7'
requests
setuptools>=41.0.0
pandas==0.25.1
tensorflow==2.2.0
ffmpeg-python
norbert==0.2.1
librosa>=0.7.2
matplotlib

#updated requirements
webbrowser
pyautogui
pyperclip
pafy
bs4
requests
pytube3